Rename ShippingCountriesForm to EstimatedShippingRatesCard, and move files.
