+++ js/src/components/flat-shipping-rates-input-cards/estimated-shipping-rates-card/add-rate-modal.js
@@ -113,6 +113,6 @@ const AddRateModal = ( { countries, onRequestClose, onSubmit } ) => {
 
 export default AddRateModal;
 /**
- * @typedef {import("./countries-form.js").AggregatedShippingRate} AggregatedShippingRate
+ * @typedef {import("./estimated-shipping-rates-card.js").AggregatedShippingRate} AggregatedShippingRate
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
  */
