+++ js/src/components/flat-shipping-rates-input-cards/estimated-shipping-rates-card/edit-rate-modal.js
@@ -148,6 +148,6 @@ const EditRateModal = ( {
 export default EditRateModal;
 
 /**
- * @typedef {import("../countries-form.js").AggregatedShippingRate} AggregatedShippingRate
+ * @typedef {import("./estimated-shipping-rates-card.js").AggregatedShippingRate} AggregatedShippingRate
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
  */
