+++ js/src/components/flat-shipping-rates-input-cards/estimated-shipping-rates-card/index.js
@@ -1,0 +1,1 @@
+export { default } from './estimated-shipping-rates-card';
