+++ js/src/components/flat-shipping-rates-input-cards/estimated-shipping-rates-card/shipping-rate-input-control.scss
